+++ src/app/models/agreement.ts
@@ -1,0 +1,11 @@
+import { Permission } from "./permission";
+
+export interface Agreement { 
+    '@id': string;
+    assigner: string ;
+    assignee: string ;
+    target: string ;
+    timestamp: string ;
+    permission: [Permission] ;
+    currentCount: number
+}